fix: refactored code so that components and libs are in a consistent spot
